+++ src/components/index.js
@@ -1,6 +1,8 @@
 export { default as Button } from './Button/Button';
 export { default as Alert } from './Alert/Alert';
 export { default as Input } from './Input/Input';
+export { default as Select } from './Select/Select';
+export { default as Spinner } from './Spinner/Spinner';
 
 // Legacy components