+++ CHANGELOG.md
@@ -4,20 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.3.0] - 01/04/2022
+### Add
+- Add Select Component
+- Add SelectPhone Component
+- Add Spinner Component
 
 ## [0.2.1] - 25/03/2022
 ### Fix
 - GlobalStyles duplicated
 
 ## [0.2.0] - 24/03/2022
-### add
+### Add
 - Add Input Component
-### update
+### Update
 - Tailwind css v3
 ### Delete
 - @apprecio/apprecio-tailwindcss-util library (deprecated)
 
 ## [0.1.0] - 15/02/2022
-### add
+### Add
 - configuration project with styleguidist, rollup and cra4
 - Add Button Component